requirements.txt
asn1crypto==0.24.0
certifi==2020.4.5.1
chardet==3.0.4
cryptography==2.1.4
Deprecated==1.2.9
enum34==1.1.6
idna==2.9
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
PyGithub==1.45
pygobject==3.26.1
PyJWT==1.7.1
pyxdg==0.25
requests==2.23.0
SecretStorage==2.3.1
six==1.14.0
urllib3==1.25.9
wrapt==1.12.1
